+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 2026-08-26
-- Implement a GroupByTerm class.  Experimental for now. 
+- Implement a GroupTerm class.  Experimental for now. 
 
 ## 2026-02-22 (version 3.0.5)
 - Bump up dependencies